refactor(be): move test files to test dir
